* add company success,error alerts * add checks to not delete metaland company * update loader component for form request * created success and alert component

+++ src/actions/companies.ts
@@ -2,32 +2,43 @@
 import dbConnect from "@/lib/mongodb";
 import  Company  from "@/models/companies"
 import { revalidatePath } from "next/cache";
+import { ServerActionResponse } from "@/types/server-action-reply";
 
 export const getCompanies = async() => {
     return await Company.find({})
 }
 
-export const saveCompanyAction = async(form: any) => {
+export const saveCompanyAction = async(form: any) : Promise<ServerActionResponse> => {
     await dbConnect()
     try {
         const newCompany = await Company.create(form)
-        revalidatePath("/companies");
-        // redirect('/companies')
-        return newCompany.toString();
-    } catch (error) {
-        console.dir(error);
-        return {message: 'error creating company'};
+        revalidatePath("/");
+        return {success: true, message: 'Company '+ newCompany.name + ' created', document: newCompany}
+    } catch (e: any) {
+        console.dir(e);
+        let errors = []
+        for(let field in e.errors) {
+            errors.push(e.errors[field].message)
+        }
+        return {success: false, message: 'Error creating company ', document: null, errors : errors}
     }
 }
 
-export const deleteCompanyAction = async(id: any) => {
+export const deleteCompanyAction = async(id: any) : Promise <ServerActionResponse> => {
     await dbConnect()
     try {
-        await Company.deleteOne({_id: id})
-        revalidatePath("/");
-        return ('company deleted');
-    } catch (error) {
-        console.dir(error);
-        return {message: 'error deleting company'};
+        if(id !== "66da989213b8e3d6f81346d3") {
+            await Company.deleteOne({_id: id})
+            revalidatePath("/");
+            return {success: true, message: 'company deleted', document: null}
+        }
+        return {success: false, message: 'Company cannot be deleted', document: null}
+    } catch (e:any) {
+        console.dir(e);
+        let errors = []
+        for(let field in e.errors) {
+            errors.push(e.errors[field].message)
+        }
+        return {success: false, message: 'Error in deleting company', document: null, errors: errors}
     }
 }
